+++ src/js/abstract/hocs/lifecycle-logger.js
@@ -1,0 +1,5 @@
+import maybe from '../../maybe';
+
+const lifecycleLogger = maybe((...args) => console.log(...args))()(true);
+
+export default lifecycleLogger;
